Revert "Connect using UUIDs not hardcoded names"
Summary:
This reverts commit 302118a1ac5491db3ba567baea6ffacd10d3caef.

For SOME reason, on SOME devices, we get an ArrayIndexOutOfBoundsException
in includesPrimaryServices(). The length of the array is 62, as expected,
but the index is -67. There is no reason for the index to be -67, as we are
never subtracting from it.

Temporary solution: reverting the commit so the customers can connect,
until we get a Galaxy Note 3 to test this freak.

Test Plan: as usual

Reviewers: samit.vaidya

Differential Revision: https://office.tobyrich.com/D98